Move Layout Catalog to components folder
