+++ Week-1/src/App.js
@@ -1,5 +1,5 @@
 import React, { Fragment } from 'react'
-import Layout from '~/src/Layout.js'
+import Layout from '~/src/components/Layout.js'
 import { BrowserRouter as Router } from 'react-router-dom'
 
 const App = () => (
